pyproject.toml
[project]
name = "git-ticket-reference-test"
version = "1.1.3"
description = "Pre-commit hooks om automatisch ticketreferenties toe te voegen en te valideren in Git commit-berichten."
authors = [
    {name = "Collin Aldaibis", email = "caldabis@gmail.com"},
]
dependencies = [
    "requests",
    "python-dotenv",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
prepare-commit-msg-ticket = "git_ticket_hooks.prepare_commit_message:main"
validate-commit-msg-ticket = "git_ticket_hooks.validate_commit_message:main"

[tool.pdm]
distribution = false
